refactor: address review feedback on orphan cleanup
- Move isOwnedByRoleSet to utils.go for shared use across syncers
- Aggregate deletion errors with utilerrors.NewAggregate instead of
  returning on first error, reducing reconcile cycles on transient failures
